Per-player localization
This change allows for Bolt to display messages to the user in the
language they have picked in the settings of their Minecraft client.

This is controlled by the new `per-player-locale` config option. The
`language` config option is still used, even if per-player locales are
enabled. It will be used as the fallback language if a player is using a
language the server does not have a localization for. It is also the
language used for the console.

Additionally, multiple custom localization files can be provided. Any
properties files placed in the plugin's data folder will all be loaded,
with custom translations taking precedence over built-in ones.