Prevent claimed blocks from being destroyed
BlockDestroyEvent is called when a block is about to be destroyed. We
use it here to prevent a protected block from becoming destroyed
involuntarily, for example, because it is a door and its support was
removed. For other cases, like just breaking the door itself, it will go
through other events and be handled properly

Handling BlockDestroyEvent is a last resort, it means the block would be
destroyed by an action not able to be handled through any other event.

This fix avoids ghost claims.