Further improve bolt admin find
This moves the find command results to use the new Time utility class
added in #123. This provides more readable relative time, and also
provides the exact timestamp on hover.

This change also modifies the "click to teleport" to be more obvious by
adding a hover event, and also expands the clickable area to the whole
line.

This change also patches a bug where click teleports would not work for
built-in world. This is because Bukkit uses names different from the
vanilla game. For example, what Bukkit calls "world", minecraft calls
"overworld". Thus, a different method is needed to obtain the format
that works for minecraft commands.