Fix bundle inventory interactions not working properly
Interactions with bundles have new InventoryAction enum entries in Paper
(https://github.com/PaperMC/Paper/pull/11902). Due to this, Bolt was not
properly handling these interaction and always denied them. This meant
that you could not use bundles to withdraw or deposit items even from
containers you actually have access to.

This PR fixes these interactions. Due to this being recent API and in
Paper only, EnumUtil has to be used and so it looks a bit messy.