Fix block protection type mismatch with waxing
When waxing or unwaxing a copper block, the entire type of the block
changes. This causes the block stored in bolt to desync from what the
actual block is. This causes no issues normally, but causes the
protection to be deleted when running `/bolt admin cleanup`.

This patch fixes such desyncs.