Split BlockBreakEvent listener into two
Perform the cancellation normally but the actual protection removal on
MONITOR priority. This prevents the protection from being removed, but
the block remaining, in case another plugin with HIGH or HIGHEST
priority also cancels the break event. With this patch, the protection
is removed only if the block was also sucessfully destroyed.