Don't make the player click if no modifications were made
`/bolt modify` allows specifying multiple sources, but if all of them
are invalid, for example, players that don't exist, it still makes you
click on a protection, even though doing so will make no actual changes
to the protection.

This commit prevents that, and also resolves a race condition with the
futures returned by the source transformer, most likely when looking up
users who haven't joined the server and aren't cached.