Prelimiary support for shelves
There's no way yet to know which slot is being clicked so it's checks
both deposit and withdraw for now. It's also a little weird in that you
can have connected shelves of different owners and technically modify
ones you cannot access by swapping with the shelf you can access. Not
sure how to handle this yet.

Shelves are locked by default.