Don't allow using someone else's connected shelf
Shelves are complicated, because they can act like connected double
chests, but can disconnect and reconnect at any moment with a redstone
signal. Since determining the cause of such signals is impossible,
using a regular matcher is basically impossible.

This patch instead adds special logic when interacting with shelves
that tries to find all connected shelves and check if the player could
interact with all of them. Only then is the interaction allowed.