Don't cancel BlockMultiPlaceEvent for shelves
I'm not exactly sure what this event is meant to prevent, but for some
reason it is fired if placing a shelf that would immediately connect
to other shelves. Cancelling it always means it's annoying to place
together connected shelves if they are already powered. You'd have to
unpower them before placing the shelves.

This patch adds a destroy permission check to make sure it's your own
shelves you are connecting to.