Fix replaceable interaction with matchers
Bolt checks if a block is replaceable based on the replaced block state,
but runs matchers based on the new block state. This can cause
situations where Bolt matches a chest protection, assuming the replaced
block is a chest, while also thinking its a replaceable block, thus
cancelling the place.

This patch checks that the replaced block is of the same type as the
protection. This is because replaceable blocks, if protected, would only
match themselves, and they cannot support other blocks.